refactor: zookeeper, move SD structs from config to zookeeper discovery package
